efl: use compile-time chec for type of va_list
Using AC_TRY_RUN is bad for crosscompiling since you can't run the
binary you compiled (if you are not using a canadian crosstoolchain).
In this case the check is omitted and HAVE_VA_LIST_AS_ARRAY remains
undefined, which is wrong if you are compiling in a x86_64 for an ARM
architecture.



SVN revision: 82000