edbus: Fix connection name leak
Hack to fix the connection name leak. The problem is that each signal
handler has a reference to the FDO_BUS connection name and when the
penultimate connection name is released (and therefore its signal
handler), FDO_BUS is released, too.

Check if cn is still in the hash before trying to free it.

Patch by: Ulisses Furquim <ulisses@profusion.mobi>



SVN revision: 82077