efl/edje: conditional usage of fixed-point, program/calc caches.
Weirdly enough raster turned PROGRAM cache on to fix a bug saying that
the other path was not tested, but from IN-EFL/edje/configure.ac CALC
cache was enabled and PROGRAM was disabled.

I kept the PROGRAM on and enabled CALC. Added fixed point as well.

Cedric, Raster: should we keep PROGRAM and CALC cache options if the
other path is not tested? Can't we assume they are always on? How much
memory these cache add? Can't this be reduced?



SVN revision: 82410