efl/efreet: remove clearenv usage
This improves the tests on freebsd, which doesn't have clearenv()  (
though this commit was untested on such environement). clearenv was
rejected from POSIX so we better aim to stop its usage. However since
tests use setenv() this may lead to leaks, but it's better a leak than a
failing test.

On Linux all tests that were passing continue to pass.



SVN revision: 82779