efl: use autotools testsuite for running tests
Instead of just making our own "check-local" and calling the binaries
ourselves, just append them into "TESTS" variable. Then they run after
all check_PROGRAMS are compiled.

The reasons for changing are:
  1) If we change the test and call "make check" the test is not
  compiled again -- and the only way to compile it is to "make clean".
  2) There's no need to reinvent the wheel here.

With a recent version of Automake, the test output is redirected to log
files. This is good but unexpected for whom was used to the previous
way. So, be warned.



SVN revision: 82841