efl/eina tests: Allow libcheck 0.9.8 usage.
This commit is for all the poor souls that see no need to update their system
each day with everything the world may offer just to be sure they have the
latest. This commit is for people that like the idea of having something
working when it did so the day before. Kudos fot Tom for pointing out the
function.

SVN revision: 82980