efl/edbus: Fix edbus_name_owner_changed() when bus name is not on bus
If we pass the last argument as TRUE, that means user want to know the actual
bus id of the bus name and if the bus name is not registered it never notify
the user.

This bug was insert when fixing another one, because of that there more code
here to fix the previous bug too.

Patch by: José Roberto de Souza <zezsouza@gmail.com>



SVN revision: 83082