try to simplify internal library dependency.
Just use the lib/name/libname.la as libtool should be responsible to
emit dependencies to compiler when it evaluates.

This should reduce over-linking, also reducing the compile lines in
our verbose builds ;-)

NOTE: this seems to work on Fedora 18 (which also bitch about DSO), so
hopefully works on Debian and Ubuntu (and elsewhere).

Please revert if breaks builds!



SVN revision: 83105