elementary/widget - revised orientation mode code.
now if applying the orientation style is failed, it tries to apply the original style. if it fails again then default theme.

also it fixed a logic error when theme changed is happened.

It will try to apply the original style if the orientation style is invalid on theme changing.




SVN revision: 84082