evas/wayland_shm: Add evas_swapper_reconfigure().
This can be used to reconfigure a swapper to another size, without the
need to destroy the swapper itself.

Although the shm pool is not being reused even when reconfiguring to a
smaller size, it could easily be.

This change is done right now only to keep the dx and dy offsets of a
previously requested swapper, which were not still used.