Elementary elm_glview: call surface update when mode changes
When glview mode was set using elm_glview_mode_set() function, it
wasn't properly calling the surface_update function internally.
This can potentially cause the surface to be not updated to the latest
mode that was set by the user.  Also removed a call to the surface
update function that was unnecessary.