Revert "Add a new API to edbus to let it create an EDbus session from"
This reverts commit 8ecd30d578ebac46bbdf5f6d5c0b7cad1187f84f.

First of all, if it's not tested it shouldn't be committed. Seriously,
test your stuff before committing, particularly when you *add* an API.
If it slips in and we forget it there, once it's released we can no
longer remove it.

Second, EDBus is a complete wrapper over libdbus and is in our intention
to remove libdbus when the EDBus implements the wire protocol itself.
Therefore there's no point in adding a public function to create an
EDBus_Connection from a DBusConnection, even if it's passed as a void
pointer.