m4: With efl with binary macro to help cross-compilation
Stumbled over it while trying to give configure a working edje_cc
when doing cross-compile. The path was picked up but never set as
the Makefile_Edje_Helper.am guarded it with HAVE_EDJE_CC which we
never successfully assigned due to this typo.

Thanks goes to Daniel for another round of pair-bug-spotting.