eeze/sensor/tizen: Fix bug where we set the wrong sensor type to the object.
Eeze sensor and Tizen types can't be in sync as the order have been changed
several times in Tizen already. We need to keep the order stable for ABI
compliance though after 1.8.

While we did the translation in one direction we did not in the other. Fix
this bug that shows now.