eeze/sensor: Move eeze_Sernsor_obj_get into private header.
Together with eeze_sensor_module_register/unregister these are the functions
that are irrelevant for developer when they just want to use the eeze_sensor
API.

Still all three need to be marked with EAPI as they are used in the backend
modules and due to our visibility hidden usage the module loading would fail
if we remove EAPI from them.

Moving them into the private header should give an indication that these are
not meant for pure API users.