ecore_evas/wayland: Put the default frame object in a lower layer.
The frame should not stay in the same layer as the other objects. If the
application wants to allow it to be on top of the content, then it must
provide a custom frame object and set it to that layer.

NOTE: Should we make Elementary's be on a lower layer too?