Elementary: Add include to Efl_Config.h.
This is the first commit of the headers split phase in elementary.
Now, each widget will have 2 or 3 h files:
- widget_Eo.h: Eo API functions (functions defines, enums, base id).
- widget_Legacy.h: contains the API functions related to objects
- widget_Common.h: common data (structs, enums...) + functions not related to
objects. This file will exist only if needed.

This phase is needed for the EFL 1.8 release to disable Eo APIs if we
consider it is not enough mature to be used by applications.