edje: fix build break on Windows.
NOTE: In C, you can only mark an area const if it is not changed at
runtime, this include also the job of the linker at runtime. So this
array can't be const because they use pointer from Evas that will be
resolved during link at runtime, instead of link at compile time.
This pass on Linux, but fail on other architecture and is actually
not correct. So let's not use const here.