ecore: add upower system module.
It's always enabled as it's a dbus module and links to nothing,
actually the daemon doesn't need to be running -- in that case it will
do nothing. In the case the daemon becomes active then it will get the
OnLowBattery property and keep it in sync.

NOTE: I couldn't test the property change as my laptop takes many
hours to get to that situation... let's hope it works :-)