configure: relax xkbcommon version requirement for ecore-wayland
Summary:
Allow ecore-wayland to be configured and compiled with xkbcommon 0.3.0.
Ecore-wayland does not use any of the new APIs in 0.3.1 nor is it exposed
to the bug that was fixed in 0.3.1.

Most distros don't include xkbcommon > 0.3.0 yet.  Thus, if 0.3.1 is
required right now, then everyone is forced to build xkbcommon, too,
which contributes to dependency *madness*.  Of course, anyone is still
welcome to build and link to xkbcommon 0.3.1 at will.

Signed-off-by: U. Artie Eoff <ullysses.a.eoff@intel.com>

Reviewers: devilhorns, antognolli

Reviewed By: devilhorns

Differential Revision: https://phab.enlightenment.org/D203