mailmaip: Add .mailmap file to fixup various git author entries.
This is an attempt to unify the various spellings and email addresses people
use in the git author line. This simple matching rules clean these up and show
them as one single author. Which in turn makes statistics with e.g. git
shortlog much nicer.

I only fixed these up for the top 50 committer in efl. If your entry is missing
or wrong feel free to add or change it. This changes alone brings down the
unique author count to 275 from 311.